chore: create unit & cypress tests for 3 components in text-input package; modify global nesting-test
Cypress tests created for TextInput and Textarea components.
Jest/Testing-Library tests created for TextInput and TextInputWithButton
components.
The general GEL-repo-wide nesting-test.js unit test has had the
insertion of a string literal child into the component undergoing
testing removed from the test. The test previously tested for the
existence of the string literal as a text node which would fail for
elements that did not accept children such as button and input.